backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
google-generativeai
requests
fpdf
python-dotenv
email-validator
